Refact: Export the overlay markup in the separate component
